requirements.txt
streamlit
pymupdf
faiss-cpu
sentence-transformers
ibm-watsonx-ai
python-dotenv
numpy
